Updating TKViewer to stream data types from DAT files - MapRenderer is still WIP

+++ src/main/java/com/gamemode/tkviewer/file_handlers/DatFileHandler.java
@@ -131,7 +131,7 @@ public class DatFileHandler extends FileHandler {
         long currentPosition = this.filePosition;
         int length = 0;
         while(true) {
-            byte b = this.readSignedByte();
+            byte b = (byte)this.readSignedByte();
             if (b != 0) {
                 length++;
             } else {
